Revision de BD

+++ GestionTiendaRafael/src/dao/interfaces/AdministradorDAO.java
@@ -7,7 +7,7 @@ import modelo.entidades.Administrador;
 public interface AdministradorDAO {
     boolean agregarAdministrador(Administrador administrador);
     boolean actualizarAdministrador(Administrador administrador);
-    boolean eliminarAdministrador(long idAdministrador);
-    Administrador obtenerAdministradorPorId(long idAdministrador);
+    boolean eliminarAdministrador(int idAdministrador);
+    Administrador obtenerAdministradorPorId(int idAdministrador);
     List<Administrador> obtenerTodosLosAdministradores();
 }
