+++ GestionTiendaRafael/src/dao/interfaces/EmpleadoDAO.java
@@ -8,8 +8,8 @@ import modelo.enums.RolEmpleado;
 public interface EmpleadoDAO {
     boolean agregarEmpleado(Empleado empleado);
     boolean actualizarEmpleado(Empleado empleado);
-    boolean eliminarEmpleado(long idEmpleado);
-    Empleado obtenerEmpleadoPorId(long idEmpleado);
+    boolean eliminarEmpleado(int idEmpleado);
+    Empleado obtenerEmpleadoPorId(int idEmpleado);
     List<Empleado> obtenerTodosLosEmpleados();
     List<Empleado> obtenerEmpleadosPorRol(RolEmpleado rol);
 }
